feat: Enhance nano firmware with temperature and humidity sensor integration, emergency mode handling, and improved UART command processing
- Added DHT sensor support for temperature and humidity readings.
- Introduced emergency mode to halt all motors and servos.
- Updated UART command handling to include new commands for temperature, humidity, and emergency state.
- Refactored RPM calculations to store values in the sensor state structure.
- Improved serial terminal for non-blocking input on Windows.
- Added logging for RX/TX messages in the serial terminal.
- Updated build log to reflect successful builds for both ESP32 and Arduino Nano.

+++ include/ultrasonicSensor.h
@@ -4,8 +4,8 @@
 
 class UltrasonicSensor {
   private:
-    int trigPin;
-    int echoPin;
+    byte trigPin;
+    byte echoPin;
 
   public:
     // Constructor default
